refactor(crud-bench): inline redis-family batch helpers into each backend
Drop `src/util/redis_batch.rs` and inline the SET/GET/DEL pipeline logic
directly into the `BenchmarkClient` impls for `redis`, `keydb`, and
`dragonfly`. The shared module hid identical-shaped code behind one extra
indirection and forced an `iter.collect()` into a `Vec` purely so the
helper signatures could be `Vec<...>`.

Inlining pushes directly from the iterator into `redis::pipe()`, avoiding
the per-batch allocation, and keeps each backend's wire commands visible
where the rest of its CRUD lives. `batch_update_*` defers to
`batch_create_*` per backend (Redis `SET` already overwrites).
