+++ src/util/mod.rs
@@ -4,9 +4,6 @@ use std::time::Duration;
 
 pub(crate) mod sql;
 
-#[cfg(any(feature = "redis", feature = "keydb", feature = "dragonfly"))]
-pub(crate) mod redis_batch;
-
 pub(crate) fn format_duration(duration: Duration) -> String {
 	let secs = duration.as_secs();
 	if secs >= 86400 {
